Split features/pwa: SW-update to src/pwa sibling, InstallButton to components
features/pwa held two layers that weren't features:
- SW-update machinery (index.ts + update-handlers.js) never reads a store or
  renders UI — it's a delivery layer. Moved to a new src/pwa sibling of
  push/realtime/storage, with a matching boundaries element + rule (may import
  pwa, shared, lib, infra, auth). infra was rejected because it can't import
  the @shared single-flight helper.
- InstallButton.tsx is dumb UI (props/events, no store/feature imports) →
  src/components.

Repoint main.tsx, MainContent.tsx, .fallowrc.json. Full `vp check` passes;
pre-commit skipped (staged-subset tsc false-positives on lucide-solid, which
resolves fine in the full build).
