+++ src/components/InstallButton.tsx
